backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
google-genai   # for Gemini API (Reading)
groq          # for Groq API (Writing, Speaking, Listening)
